Don't delete DataSets if they have been used in an Analysis

+++ refinery/core/models.py
@@ -610,20 +610,26 @@ def _dataset_delete(sender, instance, *args, **kwargs):
     delete_data_set_neo4j(instance.uuid)
 
     # Delete NodeCollection and related objs. based on uuid of
-    # Investigations
-    # This may delete any associated Studys, Assays, Nodes, Annotated
-    # Nodes??????? in addition to the related objects detected by Django
+    # Investigations. This deletes Studys, Assays and Investigations in
+    # addition to the related objects detected by Django
 
-    related_investigation_links = InvestigationLink.objects.filter(
-                                                            data_set=instance)
-    if related_investigation_links:
-        for item in related_investigation_links:
-            node_collection = NodeCollection.objects.get(
-                uuid=item.investigation.uuid)
-            try:
-                node_collection.delete()
-            except Exception as e:
-                logger.debug("Couldn't delete NodeCollection:", e)
+    if len(Analysis.objects.filter(data_set=instance)) != 0:
+        logger.error("Cannot delete DataSet:%s because there is has been "
+                     "one or more Analyses run on it." % instance)
+        raise forms.ValidationError(
+            "Cannot delete DataSet:%s because there is has been "
+            "one or more Analyses run on it." % instance)
+    else:
+        related_investigation_links = InvestigationLink.objects.filter(
+            data_set=instance)
+        if related_investigation_links:
+            for item in related_investigation_links:
+                node_collection = NodeCollection.objects.get(
+                    uuid=item.investigation.uuid)
+                try:
+                    node_collection.delete()
+                except Exception as e:
+                    logger.debug("Couldn't delete NodeCollection:", e)
 
 
 class InvestigationLink(models.Model):
